+++ src/main/kotlin/fr/fabienhebuterne/marketplace/storage/ListingsRepository.kt
@@ -5,6 +5,6 @@ import org.bukkit.inventory.ItemStack
 import java.util.*
 
 interface ListingsRepository : PaginationRepository<Listings> {
-    fun find(sellerUuid: UUID, itemStack: ItemStack, price: Long): Listings?
+    fun find(sellerUuid: UUID, itemStack: ItemStack, price: Double): Listings?
     fun findByUUID(sellerUuid: UUID): List<Listings>
 }
